REFACTOR `SocialTest` to root 'Tests' namespace
